Initial commit of JMraidcon. Don't expect anything real to work just yet
The communication is working, but only fixed commands can be sent out
and the responses are only printed in Hex/ASCII, no parsing done.

Also the sector all communication is written to is hard-coded, the
program should search for an empty sector to make sure no data gets lots
if the program crashes.